redash/api/requirements.txt
langchain==0.1.19
langchain-openai==0.1.6
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
openai==1.28.0
python-dotenv==0.19.2
quart==0.19.5
